+++ mc2/controllers/base/urls.py
@@ -32,6 +32,9 @@ urlpatterns = patterns(
     url(
         r'^exists/(?P<controller_pk>\d+)/$',
         views.update_marathon_exists_json,
-        name='update_marathon_exists_json'
-    )
+        name='update_marathon_exists_json'),
+    url(
+        r'^restarthook/(?P<controller_pk>\d+)/(?P<token>[\w-]+)/$',
+        views.ControllerWebhookRestartView.as_view(),
+        name='webhook_restart'),
 )