+++ mc2/controllers/docker/forms.py
@@ -10,12 +10,16 @@ class DockerControllerForm(ControllerForm):
     docker_image = forms.CharField(
         widget=forms.TextInput(attrs={'class': 'form-control'}))
     marathon_health_check_path = forms.CharField(
-        widget=forms.TextInput(attrs={'class': 'form-control'}),
+        widget=forms.TextInput(attrs={
+            'class': 'form-control',
+            'placeholder': '(optional)'}),
         required=False)
     port = forms.CharField(
         widget=forms.NumberInput(attrs={'class': 'form-control'}))
     domain_urls = forms.CharField(
-        widget=forms.TextInput(attrs={'class': 'form-control'}),
+        widget=forms.TextInput(attrs={
+            'class': 'form-control',
+            'placeholder': '(optional)'}),
         required=False)
     volume_needed = forms.BooleanField(
         required=False, label="Do you want storage?", initial=False,
@@ -23,13 +27,17 @@ class DockerControllerForm(ControllerForm):
     volume_path = forms.CharField(
         widget=forms.TextInput(attrs={'class': 'form-control'}),
         required=False)
+    webhook_token = forms.UUIDField(
+        required=False,
+        widget=forms.TextInput(attrs={'class': 'form-control'}))
 
     class Meta:
         model = DockerController
         fields = (
             'name', 'marathon_cpus', 'marathon_mem', 'marathon_instances',
             'marathon_cmd', 'docker_image', 'marathon_health_check_path',
-            'port', 'domain_urls', 'volume_needed', 'volume_path')
+            'port', 'domain_urls', 'volume_needed', 'volume_path',
+            'webhook_token', 'description')
 
 
 class DockerControllerFormHelper(ControllerFormHelper):