+++ mc2/controllers/docker/hidden_import.py
@@ -12,8 +12,10 @@ from django.contrib import messages
 from django.http import HttpResponseRedirect
 from django.shortcuts import render
 
+from mc2.controllers.base import tasks
 from mc2.controllers.base.views import ControllerViewMixin
 from mc2.controllers.docker.models import DockerController
+from mc2.organizations.utils import active_organization
 
 
 class HiddenImportForm(forms.Form):
@@ -29,9 +31,11 @@ class HiddenImportView(ControllerViewMixin):
     def post(self, request):
         form = HiddenImportForm(request.POST)
         assert form.is_valid()
-        DockerController.from_marathon_app_data(
-            self.request.user, json.loads(form.cleaned_data['app_data']),
+        controller = DockerController.from_marathon_app_data(
+            self.request.user, active_organization(self.request),
+            json.loads(form.cleaned_data['app_data']),
             form.cleaned_data['name'])
+        tasks.start_new_controller.delay(controller.pk)
         messages.info(self.request, u"App created: {name}".format(
             **form.cleaned_data))
         return HttpResponseRedirect('/')